move tests for remove_one_matching_row_and_return_its_line_number
