+++ tests/test_removers_feature_remover.py
@@ -1,5 +1,6 @@
 import pytest
 
+from langworld_db_data.tools.files.csv_xls import remove_one_matching_row_and_return_its_line_number
 from langworld_db_data.removers.feature_remover import FeatureRemover, FeatureRemoverError
 from tests.helpers import check_existence_of_output_csv_file_and_compare_with_gold_standard
 from tests.paths import (
@@ -68,7 +69,7 @@ def dummy_rows_of_features():
 
 
 @pytest.fixture(scope="function")
-def dummy_rows_of_lised_values():
+def dummy_rows_of_listed_values():
     return (
         {
             "id": "A-1-1",
@@ -145,201 +146,9 @@ def dummy_rows_of_feature_profile():
     )
 
 
-def test__remove_one_row_and_return_its_line_number_remove_from_inventory_of_features(
-    test_remover,
-    dummy_rows_of_features,
-):
-
-    GOLD_STANDARD_DUMMY_ROWS = (
-        {
-            "id": "A-1",
-            "en": "Subject",
-            "ru": "Подлежащее",
-        },
-        {
-            "id": "A-2",
-            "en": "Object",
-            "ru": "Прямое дополнение",
-        },
-        {
-            "id": "B-1",
-            "en": "Collocation",
-            "ru": "Коллокация",
-        },
-        {
-            "id": "B-2",
-            "en": "Grammatical core",
-            "ru": "Грамматическая основа",
-        },
-        {
-            "id": "С-1",
-            "en": "Sentence",
-            "ru": "Предложение",
-        },
-    )
-
-    rows_with_one_line_removed, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
-            match_column_name="id",
-            match_content="A-3",
-            rows=dummy_rows_of_features,
-        )
-    )
-
-    assert rows_with_one_line_removed == GOLD_STANDARD_DUMMY_ROWS
-
-    assert line_number_of_removed_row == 2
-
-
-def test__remove_one_row_from_inventory_of_listed_values(test_remover, dummy_rows_of_lised_values):
-
-    GOLD_STANDARD_DUMMY_ROWS = (
-        {
-            "id": "A-1-1",
-            "feature_id": "A-1",
-            "en": "Nominative",
-            "ru": "Номинатив",
-        },
-        {
-            "id": "A-1-3",
-            "feature_id": "A-1",
-            "en": "DSM",
-            "ru": "Дифференцированное маркирование субъекта",
-        },
-        {
-            "id": "B-1-1",
-            "feature_id": "B-1",
-            "en": "Agreement",
-            "ru": "Согласование",
-        },
-        {
-            "id": "B-1-2",
-            "feature_id": "B-1",
-            "en": "Word order",
-            "ru": "Порядок слов",
-        },
-        {
-            "id": "B-2-1",
-            "feature_id": "B-2",
-            "en": "Coordination",
-            "ru": "Координация",
-        },
-    )
-
-    rows_with_one_line_removed, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
-            match_column_name="id",
-            match_content="A-1-2",
-            rows=dummy_rows_of_lised_values,
-        )
-    )
-
-    assert rows_with_one_line_removed == GOLD_STANDARD_DUMMY_ROWS
-
-    assert line_number_of_removed_row == 1
-
-
-def test__remove_one_row_from_a_feature_profile(test_remover, dummy_rows_of_feature_profile):
-
-    GOLD_STANDARD_DUMMY_ROWS = (
-        {
-            "feature_id": "A-1",
-            "feature_name_ru": "Некий признак",
-            "value_type": "listed",
-            "value_id": "A-1-1",
-        },
-        {
-            "feature_id": "A-2",
-            "feature_name_ru": "Еще один признак",
-            "value_type": "listed",
-            "value_id": "A-2-2",
-        },
-        {
-            "feature_id": "A-3",
-            "feature_name_ru": "Еще один признак",
-            "value_type": "listed",
-            "value_id": "A-3-3",
-        },
-        {
-            "feature_id": "C-1",
-            "feature_name_ru": "И еще признак",
-            "value_type": "listed",
-            "value_id": "C-1-1",
-        },
-    )
-
-    rows_with_one_line_removed, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
-            match_column_name="feature_id",
-            match_content="B-1",
-            rows=dummy_rows_of_feature_profile,
-        )
-    )
-
-    assert rows_with_one_line_removed == GOLD_STANDARD_DUMMY_ROWS
-
-    assert line_number_of_removed_row == 3
-
-
-def test__remove_one_row_remove_last_row(test_remover, dummy_rows_of_feature_profile):
-
-    GOLD_STANDARD_DUMMY_ROWS = (
-        {
-            "feature_id": "A-1",
-            "feature_name_ru": "Некий признак",
-            "value_type": "listed",
-            "value_id": "A-1-1",
-        },
-        {
-            "feature_id": "A-2",
-            "feature_name_ru": "Еще один признак",
-            "value_type": "listed",
-            "value_id": "A-2-2",
-        },
-        {
-            "feature_id": "A-3",
-            "feature_name_ru": "Еще один признак",
-            "value_type": "listed",
-            "value_id": "A-3-3",
-        },
-        {
-            "feature_id": "B-1",
-            "feature_name_ru": "Четвертый признак",
-            "value_type": "listed",
-            "value_id": "B-1-1",
-        },
-    )
-
-    rows_with_one_line_removed, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
-            match_column_name="feature_id",
-            match_content="C-1",
-            rows=dummy_rows_of_feature_profile,
-        )
-    )
-
-    assert rows_with_one_line_removed == GOLD_STANDARD_DUMMY_ROWS
-
-    assert line_number_of_removed_row == 4
-
-
-def test__remove_one_row_throws_exception_invalid_match_content(
-    test_remover, dummy_rows_of_lised_values
-):
-
-    for bad_arg in ("abc", "A-189"):
-        with pytest.raises(ValueError, match="Row with given properties not found"):
-
-            _, _ = test_remover._remove_one_matching_row_and_return_its_line_number(
-                match_column_name="id",
-                match_content=bad_arg,
-                rows=dummy_rows_of_lised_values,
-            )
-
-
 def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_remove_values_of_A_1(
     test_remover,
-    dummy_rows_of_lised_values,
+    dummy_rows_of_listed_values,
 ):
 
     GOLD_STANDARD_DUMMY_ROWS = (
@@ -366,7 +175,7 @@ def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_r
     rows_with_multiple_rows_removed, range_of_lin_numbers_of_removed_rows = (
         test_remover._remove_multiple_matching_rows_and_return_range_of_their_line_numbers(
             match_content="A-1",
-            rows=dummy_rows_of_lised_values,
+            rows=dummy_rows_of_listed_values,
         )
     )
 
@@ -377,7 +186,7 @@ def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_r
 
 def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_remove_values_of_B_1(
     test_remover,
-    dummy_rows_of_lised_values,
+    dummy_rows_of_listed_values,
 ):
 
     GOLD_STANDARD_DUMMY_ROWS = (
@@ -410,7 +219,7 @@ def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_r
     rows_with_multiple_rows_removed, range_of_lin_numbers_of_removed_rows = (
         test_remover._remove_multiple_matching_rows_and_return_range_of_their_line_numbers(
             match_content="B-1",
-            rows=dummy_rows_of_lised_values,
+            rows=dummy_rows_of_listed_values,
         )
     )
 
@@ -421,7 +230,7 @@ def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_r
 
 def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_remove_value_of_B_2(
     test_remover,
-    dummy_rows_of_lised_values,
+    dummy_rows_of_listed_values,
 ):
 
     GOLD_STANDARD_DUMMY_ROWS = (
@@ -460,7 +269,7 @@ def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_r
     rows_with_multiple_rows_removed, range_of_lin_numbers_of_removed_rows = (
         test_remover._remove_multiple_matching_rows_and_return_range_of_their_line_numbers(
             match_content="B-2",
-            rows=dummy_rows_of_lised_values,
+            rows=dummy_rows_of_listed_values,
         )
     )
 
@@ -471,13 +280,13 @@ def test__remove_multiple_matching_rows_and_return_range_of_their_line_numbers_r
 
 def test__remove_multiple_matching_rows_throws_exception_invalid_feature_id(
     test_remover,
-    dummy_rows_of_lised_values,
+    dummy_rows_of_listed_values,
 ):
 
     with pytest.raises(ValueError, match="Rows with given properties not found"):
         test_remover._remove_multiple_matching_rows_and_return_range_of_their_line_numbers(
             match_content="X-1",
-            rows=dummy_rows_of_lised_values,
+            rows=dummy_rows_of_listed_values,
         )
 
 
@@ -515,7 +324,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_features_update
     )
 
     rows_without_A_2, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
+        remove_one_matching_row_and_return_its_line_number(
             match_column_name="id",
             match_content="A-2",
             rows=dummy_rows_of_features,
@@ -570,7 +379,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_features_update
     )
 
     rows_without_last_feature_in_category_A, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
+        remove_one_matching_row_and_return_its_line_number(
             match_column_name="id",
             match_content="A-3",
             rows=dummy_rows_of_features,
@@ -596,7 +405,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_features_update
 
 def test__update_indices_after_given_line_number_if_necessary_in_listed_values_update_is_necessary(
     test_remover,
-    dummy_rows_of_lised_values,
+    dummy_rows_of_listed_values,
 ):
 
     GOLD_STANDARD_DUMMY_ROWS = (
@@ -629,7 +438,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_listed_values_u
     rows_without_B_1, range_of_line_numbers_of_removed_rows = (
         test_remover._remove_multiple_matching_rows_and_return_range_of_their_line_numbers(
             match_content="B-1",
-            rows=dummy_rows_of_lised_values,
+            rows=dummy_rows_of_listed_values,
         )
     )
 
@@ -662,7 +471,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_listed_values_u
 
 def test__update_indices_after_given_line_number_if_necessary_in_listed_values_update_is_not_necessary(
     test_remover,
-    dummy_rows_of_lised_values,
+    dummy_rows_of_listed_values,
 ):
 
     GOLD_STANDARD_DUMMY_ROWS = (
@@ -701,7 +510,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_listed_values_u
     rows_without_B_2, range_of_line_numbers_of_removed_rows = (
         test_remover._remove_multiple_matching_rows_and_return_range_of_their_line_numbers(
             match_content="B-2",
-            rows=dummy_rows_of_lised_values,
+            rows=dummy_rows_of_listed_values,
         )
     )
 
@@ -765,7 +574,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_feature_profile
     )
 
     rows_without_A_2, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
+        remove_one_matching_row_and_return_its_line_number(
             match_column_name="feature_id",
             match_content="A-2",
             rows=dummy_rows_of_feature_profile,
@@ -823,7 +632,7 @@ def test__update_indices_after_given_line_number_if_necessary_in_feature_profile
     )
 
     rows_without_last_feature_in_category_A, line_number_of_removed_row = (
-        test_remover._remove_one_matching_row_and_return_its_line_number(
+        remove_one_matching_row_and_return_its_line_number(
             match_column_name="feature_id",
             match_content="A-3",
             rows=dummy_rows_of_feature_profile,
